% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fee_ind.R
\name{communityMST}
\alias{communityMST}
\title{Calculate community's minimum spanning tree (MST)}
\usage{
communityMST(
  abund,
  spp_traits,
  abundWeighted = TRUE,
  dis_metric = c("euclidean", "manhattan")
)
}
\arguments{
\item{abund}{a numeric vector of species abundances in the community of
interest. Species order in this vector needs to be same as that in species
trait data (rows of \code{spp_traits}). Thus its length needs to be same as
the number of rows in \code{spp_traits}.}

\item{spp_traits}{a matrix (or data frame) of numeric functional traits for
species. Columns are for different traits and rows are for different
species in the species pool.}

\item{abundWeighted}{logical; indicating whether the abundance values in
\code{abund} are considered as weights in MST calculation. The default is
\code{TRUE}.}

\item{dis_metric}{string specifying the scheme of quantifying species
dissimilarity in trait space. The currently available options are
\code{"euclidean"} (the default) and \code{"manhattan"}. Euclidean
distances are root sum-of-squares of differences, and manhattan distances
are the sum of absolute differences.}
}
\value{
If \code{abundWeighted} is \code{TRUE}, return both sorted branch
  lengths of abundance-weighted MST and sum of branch lengths of raw MST. If
  \code{abundWeighted} is \code{FALSE}, return sorted branch lengths of MST.
}
\description{
Get branch length information of MST for the community specified in
\code{abund} in the trait space defined by \code{spp_traits}.
}
\details{
...
}
