% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fee_ind.R
\name{computeFEE}
\alias{computeFEE}
\title{Functional Extension and Evenness (FEE) index}
\usage{
computeFEE(
  pool_traits,
  abund,
  abundWeighted = TRUE,
  dis_metric = c("euclidean", "manhattan"),
  doFEE0 = FALSE,
  poolBased = FALSE,
  user_ecdf = NULL,
  doParallel = FALSE
)
}
\arguments{
\item{pool_traits}{a matrix (or data frame) of numeric functional traits for
species pool. Columns are for different traits and rows are for different
species in the species pool.}

\item{abund}{matrix (or data frame) of species abundances in the communities
of interest. Columns are for different species in the species pool (in the
same species order as in \code{pool_traits}. Thus its number of columns
needs to be equal to the number of rows in \code{pool_traits}). Each row
is for a community. For species not presented in a community, their
abundance values should be assigned 0 or \code{NA}.}

\item{abundWeighted}{logical; indicating whether the abundance values in
\code{abund} are considered as weights in FEE calculation. The default is
\code{TRUE}.}

\item{dis_metric}{string specifying the scheme of quantifying species
dissimilarity in trait space. The currently available options are
\code{"euclidean"} (the default) and \code{"manhattan"}. Euclidean
distances are root sum-of-squares of differences, and manhattan distances
are the sum of absolute differences.}

\item{doFEE0}{logical; indicating whether the output includes FEE0 (raw
functional extension and evenness metric). The default is \code{FALSE}.}

\item{poolBased}{logical; indicating whether the calculation is based on
species pool. The default is \code{FALSE}, indicating no prior knowledge
about trait distribution (i.e., adopting uniform distribution for trait
values to build null community models). If \code{TRUE}, \code{pool_traits}
is used to build null community models.}

\item{user_ecdf}{string specifying the path and name of the file that saves
user's results of empirical cumulative distribution function (eCDF) of
FEE0. Under the same setting (\code{pool_traits} and \code{dis_metric}),
calculated eCDFs could be reused to save time. If the parameter is
\code{NULL} (default), no user's eCDF will be adopted or saved, and the
species-pool-based eCDFs will be calculated from scratch. If the file is
given, the eCDFs contained in this file will be adopted as needed, and
when the eCDF of a specific combination of species number and trait number
is not available in the file, it will be calculated from scratch and then
saved to the file.}

\item{doParallel}{logical; indicating whether the calculation of FEE is
conducted in parallel. The default value is \code{FALSE}, i.e.,
sequentially calculating the FEE index for the communities specified in
\code{abund}.}
}
\value{
The FEE index values of the communities. If \code{doFEE0} is
  \code{TRUE}, the output includes two lists: the raw FEE0 metrics and the
  FEE indices. Specifically, FEE values for monocultures are \code{NA}.
}
\description{
Computes the FEE index for the communities specified in \code{abund}, based
on the trait data specified in \code{pool_traits}.
}
