% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fee_ind.R
\name{computeNSP}
\alias{computeNSP}
\title{Calculate species richness}
\usage{
computeNSP(abund)
}
\arguments{
\item{abund}{matrix (or data frame) of species abundances in the communities
of interest. Columns are for different species in the species pool. Each
row corresponds a community. For species not presented in a community,
their abundance values should be assigned 0 or \code{NA}.}
}
\value{
Number of species in each of the communities specified in \code{abund}.
}
\description{
Computes species richness (taxonomic diversity) for the communities specified in \code{abund}.
}
\examples{
d_comm <- rbind(comm1 = c(0:9),
                comm2 = rep(1, 10),
                comm3 = runif(10, 0, 5),
                comm4 = rchisq(10, 2))
computeNSP(d_comm)
}
