% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fee_ind.R
\name{FEE0_eCDF}
\alias{FEE0_eCDF}
\title{Create empirical cumulative distribution function of FEE0}
\usage{
FEE0_eCDF(
  num_species,
  num_traits,
  dis_metric = c("euclidean", "manhattan"),
  pool_traits = NULL
)
}
\arguments{
\item{num_species}{integer; number of species.}

\item{num_traits}{integer; number of traits (i.e., dimension of the trait
space).}

\item{dis_metric}{string specifying the scheme of quantifying species
dissimilarity in trait space. The currently available options are
\code{"euclidean"} (the default) and \code{"manhattan"}. Euclidean
distances are root sum-of-squares of differences, and manhattan distances
are the sum of absolute differences.}

\item{pool_traits}{a matrix (or data frame) of numeric functional traits for
species pool. Columns are for different traits and rows are for different
species in the species pool. If \code{pool_traits} is \code{NULL}, the
calculation is then under the assumption of no prior knowledge of trait
values (i.e., uniform distribution of trait values).}
}
\value{
The eCDF of FEE0, used to transfer FEE0 metrics to FEE values.
}
\description{
Use null model approach to create empirical cumulative distribution function
(eCDF) of FEE0.
}
\details{
...
}
