% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fee_ind.R
\name{computeNSP}
\alias{computeNSP}
\title{Species richness}
\usage{
computeNSP(abund)
}
\arguments{
\item{abund}{matrix (or data frame) of species abundances in the communities
of interest. Columns are for different species in the species pool. Each
row corresponds a community. For species not presented in a community,
their abundance values should be assigned 0 or \code{NA}.}
}
\value{
Number of species in each of the communities specified in \code{abund}.
}
\description{
Computes species richness (taxonomic diversity) for the communities specified in \code{abund}.
}
